added abstract fragment for edit/create school
